+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -239,7 +239,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KEVEEQTGEREVK77G6N0A6FPZ"
+        "01KEVXWB06A6T2SP0QBCRDA3K5"
     )
     r"""The unique identifier of the data source"""