+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -130,7 +130,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KEVEEQTCZXDGBY370FVJYNT3"
+        "01KEVXWB04PMVC65SYGRZ80AD6"
     )
     r"""The unique identifier of the data source"""