+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KEVEEQTE2MSA83D8K560QDY1"
+        "01KEVXWB05REETX07CMD189G8M"
     )
     r"""The unique identifier of the data source"""